+++ inspectit-ocelot-core/src/main/java/rocks/inspectit/ocelot/core/instrumentation/actions/GenericActionGenerator.java
@@ -1,4 +1,4 @@
-package rocks.inspectit.ocelot.core.instrumentation.genericactions;
+package rocks.inspectit.ocelot.core.instrumentation.actions;
 
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
@@ -15,6 +15,8 @@ import rocks.inspectit.ocelot.bootstrap.exposed.ObjectAttachments;
 import rocks.inspectit.ocelot.bootstrap.instrumentation.IGenericAction;
 import rocks.inspectit.ocelot.config.model.instrumentation.actions.GenericActionSettings;
 import rocks.inspectit.ocelot.config.utils.AutoboxingHelper;
+import rocks.inspectit.ocelot.core.instrumentation.actions.template.GenericActionTemplate;
+import rocks.inspectit.ocelot.core.instrumentation.actions.template.VoidGenericActionTemplate;
 import rocks.inspectit.ocelot.core.instrumentation.config.model.GenericActionConfig;
 import rocks.inspectit.ocelot.core.instrumentation.injection.ClassInjector;
 import rocks.inspectit.ocelot.core.instrumentation.injection.InjectedClass;
@@ -99,10 +101,10 @@ public class GenericActionGenerator {
             clCache = actionsCache.get(loader);
             clCache.cleanUp(); //cleanup to make sure unused InjectedClasses are released
             try {
-                String id = actionConfig.isVoid() ? VOID_GENERIC_ACTION_STRUCTURAL_ID : NON_VOID_GENERIC_ACTION_STRUCTURAL_ID;
+                String templateType = actionConfig.isVoid() ? VOID_GENERIC_ACTION_STRUCTURAL_ID : NON_VOID_GENERIC_ACTION_STRUCTURAL_ID;
                 return clCache.get(actionConfig, () ->
                         (InjectedClass<? extends IGenericAction>)
-                                classInjector.inject(id, classToUseActionOn, (className) ->
+                                classInjector.inject(templateType, classToUseActionOn, (className) ->
                                         buildGenericActionByteCode(actionConfig, loader, className)
                                 ));
             } catch (ExecutionException | ExecutionError e) {
