+++ inspectit-ocelot-core/src/test/java/rocks/inspectit/ocelot/core/instrumentation/actions/MyDummyClass.java
@@ -1,0 +1,5 @@
+package rocks.inspectit.ocelot.core.instrumentation.actions;
+
+public class MyDummyClass {
+    public static int MY_VALUE = 0;
+}
