--- inspectit-ocelot-core/src/test/java/rocks/inspectit/ocelot/core/instrumentation/genericactions/MyDummyClass.java
@@ -1,5 +1,0 @@
-package rocks.inspectit.ocelot.core.instrumentation.genericactions;
-
-public class MyDummyClass {
-    public static int MY_VALUE = 0;
-}
